+++ examples/CHANGELOG.md
@@ -1,5 +1,14 @@
 # maskera-examples
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+  - @maskera/core@0.4.1
+
 ## 0.0.4
 
 ### Patch Changes